Implementacion Service Test

+++ src/main/java/com/example/backend/repository/CompraRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.backend.model.Compra;
 
 @Repository
-public interface CompraRepository extends JpaRepository<Compra, Long>{
+public interface CompraRepository extends JpaRepository<Compra, Integer>{
 
 }
