+++ src/main/java/com/example/backend/repository/ComunaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.backend.model.Comuna;
 
 @Repository
-public interface ComunaRepository extends JpaRepository<Comuna, Long>{
+public interface ComunaRepository extends JpaRepository<Comuna, Integer>{
 
 }
