+++ src/main/java/com/example/backend/repository/MarcaRepository.java
@@ -7,6 +7,6 @@ import com.example.backend.model.Marca;
 
 
 @Repository
-public interface MarcaRepository extends JpaRepository<Marca, Long>{
+public interface MarcaRepository extends JpaRepository<Marca, Integer>{
     
 }
